+++ mrb_doc/models/browser.rb
@@ -12,4 +12,32 @@ module Browser
     # src/mruby_integration/models/browser.cpp
     nil
   end
+
+  # Sets the main loop for web exports.
+  #
+  # @example Basic usage
+  #   def main
+  #     # Game logic goes here
+  #   end
+  #
+  #   Browser.main_loop = "main"
+  #
+  # @param method [String]
+  # @return [nil]
+  # @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on the wrong platform.
+  def self.main_loop=(method)
+    # mrb_set_main_loop
+    # src/web.cpp
+    nil
+  end
+
+  # Cancels the main loop for web exports.
+  #
+  # @return [nil]
+  # @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on the wrong platform.
+  def self.cancel_main_loop
+    # mrb_cancel_main_loop
+    # src/web.cpp
+    nil
+  end
 end